master - mdrew: fixed jni casting bug in socket io

+++ quiet/src/main/java/org/quietmodem/Quiet/SocketOutputStream.java
@@ -5,7 +5,7 @@ import java.io.OutputStream;
 
 public class SocketOutputStream extends OutputStream {
     private native void nativeClose() throws IOException;
-    private native void nativeWrite(byte[] buf, int off, int len) throws IOException;
+    private native void nativeWrite(byte[] buf, long off, long len) throws IOException;
 
     private int fd;
